adaptive UI and removed unnecessary API endpoint
